common/async: move service out of detail
